src/cmd: introduce info subcommand
Introduce an `info` subcommand.

This command will fetch information about the package from AUR (given
that it exists), and print it to stdout. There's a lot of improvements
to make when displaying the information: terminal colors, default value
for an empty field and handle packages that aren't found, to name a few.