src/config: store pacman repositories
Store pacman repositories in the config.

The repositories are stored in a struct `Repository` to make it easy to
manage the repositories, since they will have different severs
(mirrors). The repositories are hardcoded for now but later they should
be read from `/etc/pacman.conf` in case customizations have been made.